+++ src/main/java/sidkbk/celemo/CelemoApplication.java
@@ -3,6 +3,7 @@ package sidkbk.celemo;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.data.mongodb.config.EnableMongoAuditing;
+import sidkbk.celemo.services.TimerService;
 
 @SpringBootApplication
 @EnableMongoAuditing
@@ -10,7 +11,7 @@ public class CelemoApplication {
 
 	public static void main(String[] args) {
 		SpringApplication.run(CelemoApplication.class, args);
-		/*TimerService.checkAuctionEndTimeTimer();*/
+		TimerService.checkAuctionEndTimeTimer();
 	}
 
 }
